Update lights.ino
Updated Christmas 2024.
Mainly to connect to new WiFi router and data.  Just had to create and update an arduino_secrets.h file.
Also took opportunity to update board firmware which corrects the WiFi connection failure.
Updated lights-out time from 21 to 22 h.
Remembered to find the TimeAlarms.h file and update the number of alarms from default 6 to 100.  This is too many but we're not tight on memory.